fixed rotations and added rotation from keys and move by mouse
the figure folows the mouse , and also it can be rotated from the keys .
there is a problame when there are both rotating and moving simultaneously performed .
changed some int to doubles it the Transforms .
fixed the rotation by z axis and they all seem to work well so far.

+++ Viewer/include/Camera.h
@@ -9,16 +9,16 @@ class Camera
 private:
 	glm::mat4x4 cTransform;
 	glm::mat4x4 projection;
-	glm::mat4x4 rotationTransform(int deg, int axis);
-	glm::mat4x4 ScaleTransform(int x_scale, int y_scale, int z_scale);
-	glm::mat4x4 TranslateTransform(int x_scale, int y_scale, int z_scale);
+	glm::mat4x4 rotationTransform(double deg, int axis);
+	glm::mat4x4 ScaleTransform(double x_scale, double y_scale, double z_scale);
+	glm::mat4x4 TranslateTransform(double x_scale, double y_scale, double z_scale);
 public:
 	Camera();
 	Camera(Camera* c);
 	~Camera();
-	glm::mat4x4 GetrotationTransform(int deg, int axis);
-	glm::mat4x4 GetScaleTransform(int x_scale, int y_scale, int z_scale);
-	glm::mat4x4 GetTranslateTransform(int x_scale, int y_scale, int z_scale);
+	glm::mat4x4 GetrotationTransform(double deg, int axis);
+	glm::mat4x4 GetScaleTransform(double x_scale, double y_scale, double z_scale);
+	glm::mat4x4 GetTranslateTransform(double x_scale, double y_scale, double z_scale);
 	glm::mat4x4 creatTransform(glm::vec3 Scale_val
 		, glm::vec3 Translate_val,glm::vec2 rotat_val);
 	glm::mat4x4 get_projection();
